requirements.txt
# Web Framework
Flask==3.0.2
Werkzeug==3.0.1
waitress==3.0.0

# Database
pymongo==4.6.1
motor==3.3.2

# Environment and Configuration
python-dotenv==1.0.1
python-decouple==3.8

# Security
Flask-Session==0.6.0
bcrypt==4.1.2
PyJWT==2.8.0

# Caching
Flask-Caching==2.1.0

# Utilities
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Visualization
plotly==5.18.0

# Logging and Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Frontend
Flask-Assets==2.1.0
cssmin==0.2.0
jsmin==3.0.1

# API and serialization
Flask-RESTful==0.3.10
marshmallow==3.20.2

# File handling
python-magic==0.4.27
Pillow==10.2.0

# AI and ML
openai==1.12.0
transformers==4.37.2 